requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.2
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
python-dotenv==1.0.1
werkzeug==3.0.1
zipp==3.17.0
